Refactor vulnfeeds conversion code into a shared directory for later CVEList reuse.
