+++ vulnfeeds/cves/testdata/TestReposFromReferences_A_CVE_with_a_cgit_repo_reference_that_does_not_work_without_transformation.yaml
@@ -1,0 +1,56 @@
+---
+version: 2
+interactions:
+    - id: 0
+      request:
+        proto: HTTP/1.1
+        proto_major: 1
+        proto_minor: 1
+        content_length: 0
+        host: git.musl-libc.org
+        form:
+            service:
+                - git-upload-pack
+        headers:
+            Accept:
+                - '*/*'
+            Host:
+                - git.musl-libc.org
+            User-Agent:
+                - go-git/5.x
+        url: https://git.musl-libc.org/git/musl/info/refs?service=git-upload-pack
+        method: GET
+      response:
+        proto: HTTP/2.0
+        proto_major: 2
+        proto_minor: 0
+        content_length: 5284
+        body: "001e# service=git-upload-pack\n0000014186373b4999bfd9a9379bc4a3ca877b1c80a2a340 HEAD\0multi_ack thin-pack side-band side-band-64k ofs-delta shallow deepen-since deepen-not deepen-relative no-progress include-tag multi_ack_detailed allow-tip-sha1-in-want allow-reachable-sha1-in-want no-done symref=HEAD:refs/heads/master object-format=sha1 agent=git/2.43.0\n003f86373b4999bfd9a9379bc4a3ca877b1c80a2a340 refs/heads/master\n003f56be2e4c010944259bfa1b52b5044b517e31c200 refs/heads/rs-1.0\n003e0b44a0315b47dd8eced9f3b7f31580cf14bbfc01 refs/tags/v0.5.0\n003e8f4865c8d25e508154a1cb89291f79489a4b3b4c refs/tags/v0.5.9\n003e982a47843316e65f2b849e9def893195396aeec0 refs/tags/v0.6.0\n003e9f50b77d2ba154ba0d8cbccea94af1f7b902fd1a refs/tags/v0.7.0\n003efc27b96c7d9b2c82679508a57ac998d773f14f08 refs/tags/v0.7.1\n003fc5025bfabefe952507cf333ea89854d6ba612549 refs/tags/v0.7.10\n003f2cee45762a1e69444cd711a1a8112a401540081a refs/tags/v0.7.11\n003f58483f0afdc440edf992ff75cb3209a44f624776 refs/tags/v0.7.12\n003ed1c24ac4ec721ea49c15b277eaebed30f8df1e4c refs/tags/v0.7.5\n003ebd1f6eba8ca41994e7b7aefbe1513bd1b24f1279 refs/tags/v0.7.6\n003e00817f07392c15d3614aa996c45767be8baca6e1 refs/tags/v0.7.7\n003e91e836fda7d4e2a4ba38d5faec6e3876b1b5e179 refs/tags/v0.7.8\n003edb9915ee97670885f3f2302d8c41ac6c61f2d77e refs/tags/v0.7.9\n003e801b311a5b6f00a9490ae9933439d1e62451fbd3 refs/tags/v0.8.0\n003ed1d141c5f128d2ea6efa0e5aad25c3f8af41cadc refs/tags/v0.8.1\n003f0bd7ac05d1c6691a7ff33c0145adcafb4f4803df refs/tags/v0.8.10\n003e455fc98389fac09d8cf7ec4cde310a5b7ca47485 refs/tags/v0.8.2\n003e4a08c85111835c00a4c007aa3a198d9474932a97 refs/tags/v0.8.3\n003e35620515a8ce0f83f0467968f01383c8f53d44dc refs/tags/v0.8.4\n003eaab33ec0a3c46ee82ef849a4c6308c9ccf7529bc refs/tags/v0.8.5\n003eb1a8e0d45465731a757ef476cf6a5cbd31d166f9 refs/tags/v0.8.6\n003eda0acc32efc85a91c4225157d22520743c6d8f8c refs/tags/v0.8.7\n003eef2c4fd91ffb0b333f7258e3d62266414d00702e refs/tags/v0.8.8\n003ec5d1adc56c22ead8caa9effcee0518a7a8ff3c06 refs/tags/v0.8.9\n003e2d3e2a7fc13296bb682b9a48b81f1c7130a94bde refs/tags/v0.9.0\n003ebd45dc9446166186cfdd5bc89ba719afa9ee7d4c refs/tags/v0.9.1\n003f7bec92e793d4b8a349796848cf43c7329b0f2ed0 refs/tags/v0.9.10\n003f6688a778b0419eab32e715f269319248edee9da5 refs/tags/v0.9.11\n003f372a948b8139c6583045b2cd91b6b91fcb2e5881 refs/tags/v0.9.12\n003fd0f0fa484c5216710936715c176f67b3781e4b71 refs/tags/v0.9.13\n003fc47e7062ddc1c3aa24db071db0f4d1dca792f325 refs/tags/v0.9.14\n003fb589fb4e2999026494fa4bced90aeca9e613f754 refs/tags/v0.9.15\n003eec820f1262a5d6331ad0fe9b56a8a84365766fd1 refs/tags/v0.9.2\n003e8a996b7e6c3cf9f852bfcd2adf2ed68bf4725754 refs/tags/v0.9.3\n003e53de960d6f12480a041b7b8cbf65cd1d3f3f9b23 refs/tags/v0.9.4\n003e7ad42b949665428e6f02c543bbc16172af5bbdeb refs/tags/v0.9.5\n003ebd261bf25e000a1b754616db0b2ee1e14835a75d refs/tags/v0.9.6\n003edd762cf6e5886ed297886a19a2e2ffd49157a786 refs/tags/v0.9.7\n003ee6dcebd5efa2d390f0a24dc11444024e9fd1990c refs/tags/v0.9.8\n003ea57ac9bf64ba202bd94d1b81bddd93b850c7b9d6 refs/tags/v0.9.9\n003e8afa7cf5d00eac6dbfca3da9d8678f32f650a00d refs/tags/v1.0.0\n003e84613196aad69fad6d1a2c1814fe0bfea67efbd7 refs/tags/v1.0.1\n003e0f8b7e0b61bc59c89f97140d263ff7fc427c1c34 refs/tags/v1.0.2\n003e30bd499ae1f62e9d2fad4282d42057083709e0eb refs/tags/v1.0.3\n003e8f514e76f62da2708fa04277d611bb4c2860ee52 refs/tags/v1.0.4\n003e56be2e4c010944259bfa1b52b5044b517e31c200 refs/tags/v1.0.5\n003eb1ef8067863c52c6af03f623dbecda47204101b1 refs/tags/v1.1.0\n003e0b4e0732db5e6ed0cca78d787cbd764248fcbaac refs/tags/v1.1.1\n003fe6f09004af4d3ffc4e4ad5b5d33cf45891c25f25 refs/tags/v1.1.10\n003f650b04f8cc66d4153fa7e47505dae6d8f0a24704 refs/tags/v1.1.11\n003f615629bd6fcd6ddb69ad762e679f088c7bd878e2 refs/tags/v1.1.12\n003f9c102700a7b9e743da692d3cb6bfa580fcdd4379 refs/tags/v1.1.13\n003f074e6b3fe3f2ad89b991f3d8ab0932035b312a62 refs/tags/v1.1.14\n003ffaf69b9a73d09fafcbe4fd3007b8d8724293d8e1 refs/tags/v1.1.15\n003f8fe1f2d79b275b7f7fb0d41c99e379357df63cd9 refs/tags/v1.1.16\n003f2cd663fb2d576d590a08c1e40386c07b378d5ad6 refs/tags/v1.1.17\n003feb03bde2f24582874cb72b56c7811bf51da0c817 refs/tags/v1.1.18\n003f55df09bfccbfe21fc9dd7d8f94550c0ff25ace04 refs/tags/v1.1.19\n003e3ff3a1d541241699f84b52e5ef4bacce2105568f refs/tags/v1.1.2\n003f0fa1e638e87cf257e9f96b4019b2076afd674a19 refs/tags/v1.1.20\n003f1691b23955590d1eb66a11158fdd91c86337e886 refs/tags/v1.1.21\n003fe97681d6f2c44bf5fa9ecdd30607cb63c780062e refs/tags/v1.1.22\n003fb07d45eb01e900f0176894fdedab62285f5cb8be refs/tags/v1.1.23\n003fea9525c8bcf6170df59364c4bcd616de1acf8703 refs/tags/v1.1.24\n003ede337c9314e4dede7b26e03f29cf768e7b0615b5 refs/tags/v1.1.3\n003e00733dd1cf791d13ff6155509cf139a5f7b2eecb refs/tags/v1.1.4\n003e60a4783ed4bd0496f7eb31f9b7eb2ad10df46677 refs/tags/v1.1.5\n003e996d148bf14b477b07fa3691bffeb930c67b2b62 refs/tags/v1.1.6\n003e8c1c57a64b664340b0ede1c1f0eff8212b475ef5 refs/tags/v1.1.7\n003e115af23942e024f78779ab29832fed3938c799e7 refs/tags/v1.1.8\n003edbf74a98be1e5c7564566c2ad30ce85dcf768df6 refs/tags/v1.1.9\n003e040c1d16b468c50c04fc94edff521f1637708328 refs/tags/v1.2.0\n003e73cc775bee53300c7cf759f37580220b18ac13d3 refs/tags/v1.2.1\n003e85e0e3519655220688e757b9d5bfd314923548bd refs/tags/v1.2.2\n003e7a43f6fea9081bdd53d8a11cef9e9fab0348c53d refs/tags/v1.2.3\n003ef5f55d6589940fd2c2188d76686efe3a530e64e0 refs/tags/v1.2.4\n003e0784374d561435f7c787a555aeab8ede699ed298 refs/tags/v1.2.5\n0000"
+        headers:
+            Accept-Ranges:
+                - bytes
+            Cache-Control:
+                - no-cache, max-age=0, must-revalidate
+            Content-Length:
+                - "5284"
+            Content-Security-Policy:
+                - default-src 'self'; style-src 'self' 'unsafe-inline';
+            Content-Type:
+                - application/x-git-upload-pack-advertisement
+            Date:
+                - Wed, 02 Jul 2025 23:43:30 GMT
+            Expires:
+                - Fri, 01 Jan 1980 00:00:00 GMT
+            Pragma:
+                - no-cache
+            Server:
+                - tipidee/0.0.4.0
+            Vary:
+                - Accept-Encoding
+            X-Content-Type-Options:
+                - nosniff
+            X-Frame-Options:
+                - DENY
+        status: 200 OK
+        code: 200
+        duration: 757.6502ms
