+++ vulnfeeds/cves/testdata/TestReposFromReferences_A_CVE_with_a_repo_not_already_present_in_the_VendorRepo_cache_(that_happens_to_have_a_useful_commit_and_a_repo_that_has_no_tags).yaml
@@ -1,0 +1,54 @@
+---
+version: 2
+interactions:
+    - id: 0
+      request:
+        proto: HTTP/1.1
+        proto_major: 1
+        proto_minor: 1
+        content_length: 0
+        host: github.com
+        form:
+            service:
+                - git-upload-pack
+        headers:
+            Accept:
+                - '*/*'
+            Host:
+                - github.com
+            User-Agent:
+                - go-git/5.x
+        url: https://github.com/saemorris/TheRadSystem/info/refs?service=git-upload-pack
+        method: GET
+      response:
+        proto: HTTP/2.0
+        proto_major: 2
+        proto_minor: 0
+        content_length: -1
+        body: "001e# service=git-upload-pack\n000001553400f79f45f288c3719bd88d1ac6522de6636268 HEAD\0multi_ack thin-pack side-band side-band-64k ofs-delta shallow deepen-since deepen-not deepen-relative no-progress include-tag multi_ack_detailed allow-tip-sha1-in-want allow-reachable-sha1-in-want no-done symref=HEAD:refs/heads/master filter object-format=sha1 agent=git/github-e744e5203bf9\n00452077856320087e46dce427bef2a70d8b95125fc9 refs/heads/dataAnalysis\n0047941f91b0995c801ae40f53adf1ec97c7e5aef76c refs/heads/database-login\n0044ca78b5d0cdefd673d66d2e10084936a39cf0eb74 refs/heads/login-stuff\n003f3400f79f45f288c3719bd88d1ac6522de6636268 refs/heads/master\n004aa2c66c0a6f135bad4dbec2b8c16736a97f433a36 refs/heads/report_generating\n003fde8dd4bdfb7d1bbe4949e223d129dff5550124b3 refs/heads/search\n003f208242152bce2d8ad35bff878305947262f207fd refs/heads/upload\n003faccd26516858a0d29d860ef077ee38166967db27 refs/pull/10/head\n003f529ad291f4e81652a36ae81cbda0ad7493d7c8c8 refs/pull/11/head\n003f4953f69e26d53d9329dc5b4cbfa746087b3f7b94 refs/pull/12/head\n003fcdb02dd887896ff5480fc117016174fc80fedfed refs/pull/13/head\n003f4e95999585b55fcec6b05e44e8982b5a1eeab15c refs/pull/14/head\n003f068f900679dd52a144767d2740b947e198c021c2 refs/pull/16/head\n003f0b4067de0530e9d360a77b0d95f2be718678023f refs/pull/17/head\n003f123db37c4420814fb7b5b7d37841ed5f5487f49a refs/pull/18/head\n003f2077856320087e46dce427bef2a70d8b95125fc9 refs/pull/19/head\n003fe974aab464a7a729aae974b2336ead2cbc81bee5 refs/pull/20/head\n003f0c3117f99492ea90acf09cb941cf1d7b1a6817e3 refs/pull/21/head\n003f116f2d02d3986f8725e742788029b1233b0a4bee refs/pull/22/head\n003ffc0914bf7e8128855702fb4c821f820d32780317 refs/pull/26/head\n003fc5f7c7affd6432b9ba38b9a135fba2af320305fd refs/pull/27/head\n003fa2c66c0a6f135bad4dbec2b8c16736a97f433a36 refs/pull/28/head\n003f8ca90f5c7907751d3f5c59aff13f0a0b2fef856f refs/pull/29/head\n003f208242152bce2d8ad35bff878305947262f207fd refs/pull/30/head\n003fd50399f217c540d53f5e02be9c1be212eb97cff8 refs/pull/31/head\n003fc4755f74858f73990286a79d293045198bbda060 refs/pull/32/head\n003f945dd938bdbbe2e0f0e5e10f3f5dd455ea24a736 refs/pull/33/head\n003fde8dd4bdfb7d1bbe4949e223d129dff5550124b3 refs/pull/34/head\n003e941f91b0995c801ae40f53adf1ec97c7e5aef76c refs/pull/7/head\n003e0ac53a5a863da19a79bea42bb9698d5a58564eb4 refs/pull/8/head\n003ed15c9e4e17f930397bfb56f7fad715c9b99f877c refs/pull/9/head\n0000"
+        headers:
+            Cache-Control:
+                - no-cache, max-age=0, must-revalidate
+            Content-Security-Policy:
+                - default-src 'none'; sandbox
+            Content-Type:
+                - application/x-git-upload-pack-advertisement
+            Date:
+                - Wed, 02 Jul 2025 23:43:27 GMT
+            Expires:
+                - Fri, 01 Jan 1980 00:00:00 GMT
+            Pragma:
+                - no-cache
+            Server:
+                - GitHub-Babel/3.0
+            Strict-Transport-Security:
+                - max-age=31536000; includeSubDomains; preload
+            Vary:
+                - Accept-Encoding
+            X-Frame-Options:
+                - DENY
+            X-Github-Request-Id:
+                - 4418:60EC4:3FC100:4E75D4:6865C41F
+        status: 200 OK
+        code: 200
+        duration: 338.355443ms
