+++ vulnfeeds/cves/testdata/Test_maybeUpdateVPRepoCache_Test_with_no_cache.yaml
@@ -1,0 +1,3 @@
+---
+version: 2
+interactions: []
