+++ vulnfeeds/cves/versions_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/osv/vulnfeeds/internal/testutils"
 	"github.com/google/osv/vulnfeeds/models"
+	"github.com/google/osv/vulnfeeds/utility"
 	"golang.org/x/exp/slices"
 )
 
@@ -1276,6 +1277,175 @@ func TestCommit(t *testing.T) {
 			}
 			if got != tt.want {
 				t.Errorf("Commit() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestReposFromReferences(t *testing.T) {
+	type args struct {
+		CVE         string
+		cache       VendorProductToRepoMap
+		vp          *VendorProduct
+		refs        []Reference
+		tagDenyList []string
+	}
+	tests := []struct {
+		name      string
+		args      args
+		wantRepos []string
+	}{
+		{
+			name: "A CVE with a repo not already present in the VendorRepo cache (that happens to have a useful commit and a repo that has no tags)",
+			args: args{
+				CVE:   "CVE-2023-0327",
+				cache: nil,
+				vp:    &VendorProduct{"theradsystem_project", "theradsystem"},
+				refs: []Reference{
+					{
+						Source: "cna@vuldb.com",
+						Tags:   []string{"Patch", "Third Party Advisory"},
+						Url:    "https://github.com/saemorris/TheRadSystem/commit/bfba26bd34af31648a11af35a0bb66f1948752a6"},
+				},
+				tagDenyList: RefTagDenyList,
+			},
+			wantRepos: []string{"https://github.com/saemorris/TheRadSystem"},
+		},
+		{
+			name: "A CVE with a useless (vulnerability researcher) repo",
+			args: args{
+				CVE:   "CVE-2025-0211",
+				cache: nil,
+				vp:    &VendorProduct{"campcodes", "school_faculty_scheduling_system"},
+				refs: []Reference{
+					{
+						Source: "cna@vuldb.com",
+						Tags:   []string{"Exploit", "Third Party Advisory"},
+						Url:    "https://github.com/shaturo1337/POCs/blob/main/LFI%20in%20School%20Faculty%20Scheduling%20System.md"},
+				},
+				tagDenyList: RefTagDenyList,
+			},
+			wantRepos: []string(nil),
+		},
+		{
+			name: "A CVE with a cgit repo reference that does not work without transformation",
+			args: args{
+				CVE:   "CVE-2025-26519",
+				cache: nil,
+				vp:    nil,
+				refs: []Reference{
+					{
+						Source: "cna@mitre.org",
+						Tags:   nil,
+						Url:    "https://git.musl-libc.org/cgit/musl/commit/?id=c47ad25ea3b484e10326f933e927c0bc8cded3da",
+					},
+				},
+				tagDenyList: RefTagDenyList,
+			},
+			wantRepos: []string{"https://git.musl-libc.org/git/musl"},
+		},
+		{
+			name: "A CVE with a valid GitHub repo that stopped working",
+			args: args{
+				CVE:   "CVE-2016-10525",
+				cache: nil,
+				vp:    nil,
+				refs: []Reference{
+					{
+						Source: "support@hackerone.com",
+						Tags:   []string{"Patch", "Third Party Advisory"},
+						Url:    "https://github.com/dwyl/hapi-auth-jwt2/issues/111",
+					},
+				},
+				tagDenyList: RefTagDenyList,
+			},
+			wantRepos: []string{"https://github.com/dwyl/hapi-auth-jwt2"},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			testutils.SetupGitVCR(t)
+			var Logger utility.LoggerWrapper
+			if gotRepos := ReposFromReferences(tt.args.CVE, tt.args.cache, tt.args.vp, tt.args.refs, tt.args.tagDenyList, Logger); !reflect.DeepEqual(gotRepos, tt.wantRepos) {
+				t.Errorf("ReposFromReferences() = %#v, want %#v", gotRepos, tt.wantRepos)
+			}
+		})
+	}
+}
+
+func Test_maybeUpdateVPRepoCache(t *testing.T) {
+	type args struct {
+		cache VendorProductToRepoMap
+		vp    *VendorProduct
+		repos []string
+	}
+	tests := []struct {
+		name      string
+		args      args
+		wantCache VendorProductToRepoMap
+	}{
+		{
+			name: "Test with no cache",
+			args: args{
+				cache: nil,
+				vp:    &VendorProduct{"avendor", "aproduct"},
+				repos: []string{"https://github.com/google/osv.dev"},
+			},
+			wantCache: nil,
+		},
+		{
+			name: "Test with an empty cache",
+			args: args{
+				cache: VendorProductToRepoMap{},
+				vp:    &VendorProduct{"avendor", "aproduct"},
+				repos: []string{"https://github.com/google/osv.dev"},
+			},
+			wantCache: VendorProductToRepoMap{
+				VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev"},
+			},
+		},
+		{
+			name: "Test with an empty cache and an unusable repo",
+			args: args{
+				cache: VendorProductToRepoMap{},
+				vp:    &VendorProduct{"avendor", "aproduct"},
+				repos: []string{"https://github.com/vendor/repo"},
+			},
+			wantCache: VendorProductToRepoMap{},
+		},
+		{
+			name: "Test with an existing cache",
+			args: args{
+				cache: VendorProductToRepoMap{
+					VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev"},
+				},
+				vp:    &VendorProduct{"avendor", "aproduct"},
+				repos: []string{"https://github.com/google/osv-scanner"},
+			},
+			wantCache: VendorProductToRepoMap{
+				VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
+			},
+		},
+		{
+			name: "Test with an empty cache adding two values",
+			args: args{
+				cache: VendorProductToRepoMap{},
+				vp:    &VendorProduct{"avendor", "aproduct"},
+				repos: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
+			},
+			wantCache: VendorProductToRepoMap{
+				VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			testutils.SetupGitVCR(t)
+			for _, repo := range tt.args.repos {
+				MaybeUpdateVPRepoCache(tt.args.cache, tt.args.vp, repo)
+			}
+			if !reflect.DeepEqual(tt.args.cache, tt.wantCache) {
+				t.Errorf("maybeUpdateVPRepoCache() have %#v, wanted %#v", tt.args.cache, tt.wantCache)
 			}
 		})
 	}
